Move theme checkbox to settings activity
