+++ app/src/main/java/fr/free/nrw/commons/SettingsActivity.java
@@ -1,13 +1,16 @@
 package fr.free.nrw.commons;
 
 import android.content.Context;
+import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.res.Configuration;
 import android.os.Bundle;
 import android.preference.ListPreference;
 import android.preference.Preference;
 import android.preference.PreferenceActivity;
+import android.preference.PreferenceManager;
 import android.support.annotation.LayoutRes;
+import android.support.v4.content.IntentCompat;
 import android.support.v7.app.AppCompatDelegate;
 import android.view.MenuInflater;
 import android.view.View;
@@ -23,10 +26,10 @@ public class SettingsActivity extends PreferenceActivity implements SharedPrefer
         getDelegate().installViewFactory();
         getDelegate().onCreate(savedInstanceState);
         //Check prefs on every activity starts
-        if (getSharedPreferences("prefs", Context.MODE_PRIVATE).getBoolean("theme", false)) {
-            setTheme(R.style.LightAppTheme);
+        if (PreferenceManager.getDefaultSharedPreferences(this).getBoolean("theme",false)) {
+            setTheme(R.style.DarkAppTheme);
         }else {
-            setTheme(R.style.DarkAppTheme); //default
+            setTheme(R.style.LightAppTheme); //default
         }
         super.onCreate(savedInstanceState);
         super.onCreate(savedInstanceState);
@@ -76,7 +79,12 @@ public class SettingsActivity extends PreferenceActivity implements SharedPrefer
 
     @Override
     public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key) {
-
+        if(key.equals("theme")){
+            //Finish current activity and tart new one with selected theme
+            Intent intent = getIntent();
+            finish();
+            startActivity(intent);
+        }
     }
 
     // All the stuff below is just to get a actionbar that says settings...
