+++ app/src/main/java/fr/free/nrw/commons/contributions/ContributionsActivity.java
@@ -125,7 +125,6 @@ public  class       ContributionsActivity
             if (mediaDetails != null && savedInstanceState.getBoolean("mediaDetailsVisible")) {
                 // Feels awful that we have to reset this manually!
                 getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-                getSupportActionBar().setDisplayShowCustomEnabled(true);
             }
         }
         requestAuthToken();
