+++ app/src/main/res/menu/fragment_contributions_list.xml
@@ -28,12 +28,6 @@
         android:title="@string/menu_nearby"
         app:showAsAction="never"
         />
-    <item android:id="@+id/menu_theme_toggle"
-        android:title="Night mode"
-        android:icon="?attr/toggleButtonIcon"
-        app:actionLayout="@layout/theme_toggle_layout"
-        app:showAsAction="never"
-        />
     <item android:id="@+id/menu_refresh"
           android:title="@string/menu_refresh"
           app:showAsAction="never"
