+++ app/src/main/res/xml/preferences.xml
@@ -19,5 +19,11 @@
             android:data="@string/beta_opt_in_link" />
     </Preference>
 
+    <CheckBoxPreference
+        android:title="@string/preference_theme"
+        android:defaultValue="false"
+        android:summary="@string/preference_theme_summary"
+        android:key="theme" />
+
 
 </PreferenceScreen>
